fix: Move all layout items out of redesign
