+++ src/layout/AuthCallback.jsx
@@ -1,4 +1,4 @@
-import Header from "../redesign/components/Header";
+import Header from "./Header";
 
 // Empty component for smoothing post-login redirects visually
 export default function AuthCallback() {
