+++ src/layout/Footer.jsx
@@ -1,5 +1,5 @@
 import style from "../style";
-import Section from "../redesign/components/Section";
+import Section from "./Section";
 import PolicyEngineMainLogo from "../images/logos/policyengine/white.svg";
 import useDisplayCategory from "../hooks/useDisplayCategory";
 import { SubscribeToPolicyEngine } from "../pages/home/HomeSubscribe";
