+++ src/layout/ShowcaseItem.jsx
@@ -1,5 +1,5 @@
-import style from "../../style";
-import useDisplayCategory from "../../hooks/useDisplayCategory";
+import style from "../style";
+import useDisplayCategory from "../hooks/useDisplayCategory";
 import EmphasisedLink from "./EmphasisedLink";
 import { FileImageOutlined } from "@ant-design/icons";
