+++ src/pages/About.jsx
@@ -1,8 +1,8 @@
-import Header from "../redesign/components/Header.jsx";
+import Header from "../layout/Header.jsx";
 import Footer from "../layout/Footer.jsx";
-import Section from "../redesign/components/Section.jsx";
+import Section from "../layout/Section.jsx";
 import style from "../style/index.js";
-import PageHeader from "../redesign/components/PageHeader.jsx";
+import PageHeader from "../layout/PageHeader.jsx";
 import { founders, staff, advisors } from "../data/staff.js";
 import useDisplayCategory from "../hooks/useDisplayCategory.jsx";
 import { Link, useLocation } from "react-router-dom";
