+++ src/pages/Jobs.jsx
@@ -1,10 +1,10 @@
 import React from "react";
-import Header from "../redesign/components/Header";
+import Header from "../layout/Header";
 import Footer from "../layout/Footer";
-import Section from "../redesign/components/Section";
+import Section from "../layout/Section";
 import style from "../style";
-import PageHeader from "../redesign/components/PageHeader";
-import { HoverBox } from "../redesign/components/HoverBox";
+import PageHeader from "../layout/PageHeader";
+import { HoverBox } from "../layout/HoverBox";
 
 function InternshipOpportunities() {
   return (
