+++ src/pages/TermsAndConditions.jsx
@@ -1,7 +1,7 @@
 import { Container } from "react-bootstrap";
 import useMobile from "../layout/Responsive";
 import { Helmet } from "react-helmet";
-import Header from "../redesign/components/Header";
+import Header from "../layout/Header";
 import Footer from "../layout/Footer";
 
 export default function TACPage() {
