+++ src/pages/Testimonials.jsx
@@ -1,6 +1,6 @@
-import Header from "../redesign/components/Header";
+import Header from "../layout/Header";
 import Footer from "../layout/Footer";
-import Section from "../redesign/components/Section";
+import Section from "../layout/Section";
 import style from "../style";
 import { fullQuotes } from "../data/quotes";
 import { allOrgs } from "data/organisations";
