+++ src/pages/UserProfilePage.jsx
@@ -1,8 +1,8 @@
 import Footer from "../layout/Footer";
-import Header from "../redesign/components/Header";
+import Header from "../layout/Header";
 import Helmet from "react-helmet";
-import Section from "../redesign/components/Section";
-import PageHeader from "../redesign/components/PageHeader";
+import Section from "../layout/Section";
+import PageHeader from "../layout/PageHeader";
 import style from "../style";
 import { Link, useParams } from "react-router-dom";
 import { useAuth0 } from "@auth0/auth0-react";
