+++ src/pages/home/HomePolicyEngineFeatures.jsx
@@ -1,9 +1,9 @@
 import style from "../../style";
-import Section from "../../redesign/components/Section";
+import Section from "../../layout/Section";
 import codingScreenshot from "../../images/home/coding_screenshot.png";
 import parameterScreenshot from "../../images/home/parameter_screenshot.png";
 import decileChartScreenshot from "../../images/home/decile_chart_screenshot.png";
-import ShowcaseItem from "../../redesign/components/ShowcaseItem";
+import ShowcaseItem from "../../layout/ShowcaseItem";
 import useCountryId from "../../hooks/useCountryId";
 
 function ImageComponent(imagePath) {
