+++ src/pages/home/HomeTransparency.jsx
@@ -1,9 +1,9 @@
 import style from "../../style";
-import Section from "../../redesign/components/Section";
+import Section from "../../layout/Section";
 import apiScreenshot from "../../images/home/api_screenshot.png";
 import githubScreenshot from "../../images/home/github_screenshot.png";
 import testingScreenshot from "../../images/home/testing_screenshot.png";
-import ShowcaseItem from "../../redesign/components/ShowcaseItem";
+import ShowcaseItem from "../../layout/ShowcaseItem";
 import useCountryId from "../../hooks/useCountryId";
 
 export default function HomeTransparency() {
