+++ src/pages/home/HomeUsedBy.jsx
@@ -1,5 +1,5 @@
 import style from "../../style";
-import Section from "../../redesign/components/Section";
+import Section from "../../layout/Section";
 import { orgData } from "data/organisations";
 import useCountryId from "../../hooks/useCountryId";
 import useDisplayCategory from "../../hooks/useDisplayCategory";
