+++ src/redesign/components/PolicyEngine.jsx
@@ -17,7 +17,7 @@ import {
 } from "../../api/call";
 import LoadingCentered from "../../layout/LoadingCentered";
 import ErrorPage from "../../layout/Error";
-import Header from "./Header";
+import Header from "../../layout/Header";
 import Testimonials from "../../pages/Testimonials";
 import CalculatorInterstitial from "../../pages/CalculatorInterstitial";
 import CitizensEconomicCouncil from "../../applets/CitizensEconomicCouncil";
